+++ MangaRipper/Forms/AboutBox.resx
