+++ MangaRipper/Forms/FormMain.resx
