+++ MangaRipper/Helpers/State.cs
@@ -1,12 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Drawing;
 using System.Windows.Forms;
 
-namespace MangaRipper.Helper
+namespace MangaRipper.Helpers
 {
     public class State
     {
